requirements.txt
tensorflow==2.16.1
pandas
dvc
mlflow==2.2.2
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
gdown
-e . # Indicates that the package should be installed in "editable" mode in current directory